Create progressbar
The difference between this and the heads up display is the progressbar. It has been used to modify the heads up display giving this a better look. By reading the comments,you realize that the heads up text has been replaced with a reference to the progressbar and that the percentage sign in the javascript file allows the progressbar to move when the user answers a question