Create loading our questions from open trivia
The API is OpenTrivia and so we make a remote API call to their database.The map gets the question and helps us transform the questions into the type we want.The formatted question formats the question to our type. Const answerchoices enables us iterate through the answers and we put them as choice 1,choice 2 and others.